# throttle threads for iterated small data tasks