--- a/src/data.table.h
+++ b/src/data.table.h
@@ -71,11 +71,6 @@
 #define NEED2UTF8(s) !(IS_ASCII(s) || (s)==NA_STRING || IS_UTF8(s))
 #define ENC2UTF8(s) (!NEED2UTF8(s) ? (s) : mkCharCE(translateCharUTF8(s), CE_UTF8))
 
-// Types used to dynamically limit DT threads when passing to num_threads
-#define OMP_BATCH 2 // limit to number of iterations (batches)
-#define OMP_ROWS 1  // limit to number of iterations (rows) divided by R_DATATABLE_ITER_PER_THREAD
-#define OMP_ALL 0   // do not limit
-
 // init.c
 extern SEXP char_integer64;
 extern SEXP char_ITime;
@@ -190,7 +185,7 @@ double wallclock();
 
 // openmp-utils.c
 void initDTthreads();
-int getDTthreads(int type, int iters);
+int getDTthreads(const int64_t n, const bool throttle);
 void avoid_openmp_hang_within_fork();
 
 // froll.c