… NEWS.md
… inst/tests/tests.Rraw
… src/fifelse.c
… src/forder.c
… src/froll.c
… src/frollR.c
… src/frolladaptive.c
… src/fsort.c
… src/gsumm.c
… src/nafill.c
… src/openmp-utils.c
… src/reorder.c
… src/subset.c
… src/types.c