Closes #3175, see https://github.com/Rdatatable/data.table/issues/3175#issuecomment-645808034

---

Closes #3205 (the 2nd part which was outstanding)

---

Closes #3438: rerun repeated uniqueN test

before patch
```
# 20 threads
   user  system elapsed 
478.054   1.334  24.300 

# 1 thread
   user  system elapsed 
  7.968   0.348   8.317 

# 40 threads
    user   system  elapsed 
1353.752    1.711   34.369 
```

after patch
```
# 20 threads
   user  system elapsed 
 10.021   0.382   8.518 

# 1 thread
   user  system elapsed 
  8.008   0.323   8.331 

# 40 threads
   user  system elapsed 
 12.542   0.376   8.279 
```

----

partially address #3735:
```r
library(data.table)
allIterations <- data.frame(v1 = runif(1e5), v2 = runif(1e5))
DoSomething <- function(row) someCalculation <- row[["v1"]] + 1
system.time(for (r in 1:nrow(allIterations)) DoSomething(allIterations[r, ]))
allIterations <- as.data.table(allIterations)
setDTthreads(1)
system.time(for (r in 1:nrow(allIterations)) DoSomething(allIterations[r, ]))
setDTthreads(40)
system.time(for (r in 1:nrow(allIterations)) DoSomething(allIterations[r, ]))
```
output before patch
```
# DF
   user  system elapsed 
  3.143   0.005   3.150 
# DT: setDTthreads(1)
   user  system elapsed 
 12.958   0.134  13.093 
# DT: setDTthreads(40)
    user   system  elapsed 
1098.068    0.747   27.851 
```
after patch
```
# DF
   user  system elapsed 
  3.129   0.020   3.149
# DT: setDTthreads(1)
   user  system elapsed 
 12.505   0.103  12.610 
# DT: setDTthreads(40)
   user  system elapsed 
 12.663   0.088  12.751 
```

----

partially address "slow uniqueN" #3739
```
library(data.table)
set.seed(1000)
# small vector, uniqueN() is slower
x <- sample(1:1e5, size = 1e2, replace = TRUE)
microbenchmark::microbenchmark(
  uniqueN(x),
  length(unique(x))
)
```
before patch
```
Unit: microseconds
              expr    min      lq     mean  median      uq     max neval
        uniqueN(x) 42.426 45.1665 55.85915 46.0855 47.1960 953.592   100
 length(unique(x))  6.498  7.7030  9.03348  8.6160  9.9365  28.824   100
```
after patch
```
Unit: microseconds
              expr    min      lq     mean  median      uq    max neval
        uniqueN(x) 14.978 15.7665 17.01930 16.2505 16.7810 77.220   100
 length(unique(x))  5.011  6.0760  6.84729  6.7140  7.3435 22.196   100
```

----

slow uniqueN by group #1120
```
library(data.table)
irisdt <- setDT(iris[sample(1:150, size = 10000, replace = TRUE), ])
irisdt[, Sepal.Width := Sepal.Width+ sample(0:50, size = 10000, replace = TRUE)]
irisdt[, Sepal.Length:= Sepal.Width+ sample(0:5000, size = 10000, replace = TRUE)]
microbenchmark::microbenchmark(
  irisdt[, uniqueN(Sepal.Width), Sepal.Length],
  irisdt[, length(unique(Sepal.Width)), Sepal.Length],
  times = 2
)
```
before patch
```
Unit: milliseconds
                                                expr      min       lq     mean   median       uq      max neval
        irisdt[, uniqueN(Sepal.Width), Sepal.Length] 343.0611 343.0611 391.6741 391.6741 440.2871 440.2871     2
 irisdt[, length(unique(Sepal.Width)), Sepal.Length] 127.9050 127.9050 128.5702 128.5702 129.2355 129.2355     2
```
after patch
```
Unit: milliseconds
                                                expr       min        lq      mean    median        uq       max neval
        irisdt[, uniqueN(Sepal.Width), Sepal.Length] 103.63772 103.63772 107.19231 107.19231 110.74689 110.74689     2
 irisdt[, length(unique(Sepal.Width)), Sepal.Length]  49.09571  49.09571  52.49689  52.49689  55.89808  55.89808     2
```