@mattdowle 2020-06-02
Related: https://github.com/Rdatatable/data.table/issues/3205#issuecomment-533333049