@jangorecki 2020-06-02
Moving logic to `getDTthreads` required to update its usage everywhere, so it touches now many files.
It also requires to review the proper type set, as it was not always obvious if `OMP_ROWS` or `OMP_BATCH` should be used.

- Using `OMP_ALL` is a backward compatible in terms of cpu utilization, and then second argument to `getDTthreads` is ignored.
- `OMP_BATCH` ignores new "iterations per thread" feature, and just limit to number of iterations (i.e. `nafill` over 2 columns doesn't need to create team of 40 threads).
- `OMP_ROWS` takes into account "iterations per thread" feature

All needs proper testing (and extending manual).