@mattdowle 2020-06-13
I just couldn't grasp the types so I changed this PR to what clicks in my head, at least. See what you think. The names OMP_ROWS and OMP_BATCH didn't click because in some cases that are not directly to do with rows, you do want to throttle. I just find it easier to see a bool flag (true/false) telling me directly whether the throttle should be applied or not in that region, so I don't have the mental hop via the type codes.
While I was changing the `getDTthreads()` calls, I also changed some of them to be throttle=false, where Jan had OMP_ROWS or OMP_ALL.
The default throttle of 1024 I just guesstimated. Roughly speaking, the throttle value means one thread for nrow<=1024, two threads for nrow<=2048, etc, subject to the same limits as before (percent of num_procs, etc).