@jangorecki 2020-06-13
PR in action, just today https://stackoverflow.com/questions/62347788/multi-threaded-data-table-in-r-much-slower-than-one-using-single-thread?noredirect=1#comment110287890_62347788