@jangorecki 2020-06-13
@mattdowle I updated timings in first post. It is interesting to see that `length(unique(` did speed up as well.