@jangorecki 2020-06-17
Just to ensure this PR address the problem on Windows platform as well.

before patch
```
# 10 threads
   user  system elapsed
 264.71  257.06  291.49

# 1 thread
   user  system elapsed
  25.14   23.88   49.01

# 20 threads
   user  system elapsed
 609.39  609.09  621.81
```
after patch
```
# 10 threads
   user  system elapsed
  22.77   23.63   46.50

# 1 thread
   user  system elapsed
  25.35   24.16   49.50

# 20 threads
   user  system elapsed
  25.81   26.43   52.75
```
Related issue is #4527 but it is about single threaded overhead, which on windows seems to be much bigger than on linux.